ui delete user

+++ backend/app/api/routes/user.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, Depends, HTTPException, status
 from sqlmodel import select
 
-from app.api.crud.user import create_user, delete_user, get_all_users, get_user
+from app.api.crud.user import create_user, delete_user, read_all_users, read_user
 from app.api.deps import CurrentUser, SessionDep, is_admin_user
 from app.models.user import User, UserCreate, UserGroup, UserPublic
 
@@ -35,12 +35,12 @@ async def create_user_api(*, user: UserCreate, session: SessionDep) -> UserPubli
     # ensure that no user with the same name or email exists
     # check whether the user already exists
     # TODO should be enforced at the database level and error should be handled
-    if get_user(username=user.username, session=session):
+    if read_user(username=user.username, session=session):
         raise HTTPException(
             status_code=status.HTTP_409_CONFLICT,
             detail="User with the same name already exists",
         )
-    if get_user(username=user.email, session=session):
+    if read_user(username=user.email, session=session):
         raise HTTPException(
             status_code=status.HTTP_409_CONFLICT,
             detail="User with the same email already exists",
@@ -73,6 +73,7 @@ async def create_user_api(*, user: UserCreate, session: SessionDep) -> UserPubli
 async def delete_user_api(
     user_id: int,
     session: SessionDep,
+    current_user: CurrentUser,
 ) -> None:
     """Delete a user. Only admin users can delete users.
 
@@ -80,6 +81,11 @@ async def delete_user_api(
         user_id(int): Identifier of the user
         session (SessionDep): Database session.
     """
+    if current_user.id == user_id and current_user.is_superuser:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="Superuser cannot delete itself",
+        )
     delete_user(user_id=user_id, session=session)
 
 
@@ -95,7 +101,7 @@ async def read_user_me(current_user: CurrentUser) -> User:
 
 
 @router.get("/{user_id}", response_model=User)
-async def read_user(
+async def read_user_api(
     user_id: int, current_user: CurrentUser, session: SessionDep
 ) -> User:
     """
@@ -110,7 +116,7 @@ async def read_user(
     if user_id == current_user.id:
         return current_user
     if current_user.is_superuser:
-        user = get_user(user_id=user_id, session=session)
+        user = read_user(user_id=user_id, session=session)
         if user:
             return user
         else:
@@ -131,4 +137,4 @@ async def read_user(
     dependencies=[Depends(is_admin_user)],
 )
 async def read_users(session: SessionDep) -> list[User]:
-    return get_all_users(session=session)
+    return read_all_users(session=session)
